fixup! Netflow: Networkview panel refactor. Performance improvement
